requirements.txt
torch>=2.0.0
rembg>=2.0.50
Pillow>=10.0.0
onnxruntime-gpu>=1.16.0
flask>=3.0.0
mediapipe>=0.10.0
opencv-python>=4.8.0
requests>=2.28.0
